Working build.
tbd_build takes a tree object and for each dir in ./ci it checks out
the tree to a tmp dir and runs the ./ci/stage/run command.

It uses tbd_run to store the output in the git object database with
pointers in the refs/builds/<tree_id>/stage directory.